requirements.txt
click>=8.0.0
openai>=1.0.0
python-dotenv>=0.19.0
pytest>=7.0.0
pytest-cov>=4.0.0
gitpython>=3.1.0
requests>=2.25.0
